Move step functions into engine
