+++ javascript/core/src/engine/createNode.ts
@@ -5,9 +5,8 @@
 import type { TableContainer } from '@datashaper/arquero'
 import type { Node } from '@datashaper/dataflow'
 
-import type { Step } from '../steps/index.js'
-import type { NodeFactory } from '../verbs/index.js'
 import * as verbs from '../verbs/index.js'
+import type { NodeFactory,Step } from './types.js'
 
 export function createNode(step: Step): Node<TableContainer> {
 	const records = verbs as any as Record<string, NodeFactory>
