+++ javascript/core/src/engine/index.ts
@@ -4,5 +4,7 @@
  */
 export * from './graph.js'
 export * from './GraphManager.js'
+export * from './readStep.js'
+export * from './types.js'
 export * from './validator.js'
 export * from './Workflow.js'
