+++ javascript/core/src/engine/readStep.ts
@@ -4,8 +4,8 @@
  */
 import type { NamedPortBinding, Verb } from '@datashaper/schema'
 
-import type { Step, StepInput } from '../steps/types.js'
 import * as defaults from '../verbs/defaults/index.js'
+import type { Step, StepInput } from './types.js'
 
 // TEMP: this creates a more readable id by doing a simple increment for each verb type
 // since this is global it will not align across pipelines or tables.
