+++ javascript/core/src/engine/types.ts
@@ -2,6 +2,8 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+import type { TableContainer } from '@datashaper/arquero'
+import type { Node } from '@datashaper/dataflow'
 import type {
 	NamedPortBinding,
 	OutputPortBinding,
@@ -73,3 +75,5 @@ export interface Step<T extends object | void | unknown = unknown> {
 		others?: NamedPortBinding[]
 	} & Record<string, NamedPortBinding>
 }
+
+export type NodeFactory = (id: string) => Node<TableContainer>
