+++ javascript/core/src/engine/validator.ts
@@ -5,7 +5,7 @@
 
 import Ajv from 'ajv'
 
-import type { WorkflowObject } from '../steps/types.js'
+import type { WorkflowObject } from './types.js'
 
 const baseUrl = 'https://microsoft.github.io/datashaper/schema/workflow'
 // TODO: change to `workflow.json` when it has been deployed to the github.io page
