+++ javascript/core/src/index.ts
@@ -4,6 +4,5 @@
  */
 export * from './engine/index.js'
 export * from './primitives.js'
-export * from './steps/index.js'
 export * from './util/index.js'
 export * from './verbs/index.js'
