--- javascript/core/src/steps/index.ts
@@ -1,6 +1,0 @@
-/*!
- * Copyright (c) Microsoft. All rights reserved.
- * Licensed under the MIT license. See LICENSE file in the project.
- */
-export * from './readStep.js'
-export * from './types.js'
