+++ javascript/core/src/util/steps.ts
@@ -6,7 +6,7 @@ import type { Verb } from '@datashaper/schema'
 import difference from 'lodash-es/difference.js'
 import intersection from 'lodash-es/intersection.js'
 
-import type { Step } from '../steps/index.js'
+import type { Step } from '../engine/index.js'
 
 enum Tags {
 	/**
