+++ javascript/core/src/verbs/index.ts
@@ -31,7 +31,6 @@ export { rollup } from './rollup.js'
 export { sample } from './sample.js'
 export { select } from './select.js'
 export { spread } from './spread.js'
-export * from './types.js'
 export { unfold } from './unfold.js'
 export { ungroup } from './ungroup.js'
 export { unhot } from './unhot.js'
